bug: 50455
Switch mailbox and mailbox group IDs back to ints.

http://bugzilla.zimbra.com/show_bug.cgi?id=50455

Copied from Perforce
 Change: 252986

+++ src/java/com/zimbra/common/util/ZimbraLog.java
@@ -620,8 +620,8 @@ public final class ZimbraLog {
     /**
      * Adds mailbox id to the current thread's logging context.
      */
-    public static void addMboxToContext(long mboxId) {
-        addToContext(C_MID, Long.toString(mboxId));
+    public static void addMboxToContext(int mboxId) {
+        addToContext(C_MID, Integer.toString(mboxId));
     }
 
     /**
